set pass and set user name

+++ Itemstore.cs
@@ -9,10 +9,10 @@ namespace final_project
     internal class Itemstore
     {
         public string Name { get; set; }
-        public decimal Price { get; set; }
+        public int Price { get; set; }
         public string ImagePath { get; set; }
 
-        public Itemstore(string name, decimal price, string imagePath)
+        public Itemstore(string name, int price, string imagePath)
         {
             Name = name;
             Price = price;
